requirements.txt
git+https://github.com/pyrogram/pyrogram
tgcrypto
pymongo[srv]==3.12.3
motor==2.5.1
marshmallow==3.14.1
umongo==3.0.1
python-dotenv
pyromod
requests
heroku3
psutil
bs4
